Move LeftMenu into the components folder
