+++ packages/core/admin/admin/src/components/LeftMenu/compos/Footer/index.js
