+++ packages/core/admin/admin/src/components/LeftMenu/compos/LinkHeader/messages.json
