+++ packages/core/admin/admin/src/components/LeftMenu/compos/index.js
@@ -1,0 +1,5 @@
+export { default as Footer } from './Footer';
+export { default as Header } from './Header';
+export { default as LinksContainer } from './Links';
+export { default as LinksSection } from './LinksSection';
+export { default as Loader } from './Loader';
