+++ packages/core/admin/admin/src/components/LeftMenu/reducer.js
