+++ packages/core/admin/admin/src/components/LeftMenu/tests/reducer.test.js
