+++ packages/core/admin/admin/src/components/LeftMenu/utils/checkPermissions.js
