+++ packages/core/admin/admin/src/components/LeftMenu/utils/generateModelsLinks.js
