+++ packages/core/admin/admin/src/components/LeftMenu/utils/getGeneralLinks.js
