+++ packages/core/admin/admin/src/components/LeftMenu/utils/getPluginSectionLinks.js
