+++ packages/core/admin/admin/src/components/LeftMenu/utils/getSettingsMenuLinksPermissions.js
