+++ packages/core/admin/admin/src/components/LeftMenu/utils/tests/checkPermissions.test.js
