+++ packages/core/admin/admin/src/components/LeftMenu/utils/tests/generateModelsLinks.test.js
