+++ packages/core/admin/admin/src/components/LeftMenu/utils/tests/getGeneralLinks.test.js
