+++ packages/core/admin/admin/src/components/LeftMenu/utils/tests/getSettingsMenuLinksPermissions.test.js
