+++ packages/core/admin/admin/src/components/LeftMenu/utils/tests/toPluginLinks.test.js
