+++ packages/core/admin/admin/src/reducers/index.js
@@ -2,7 +2,7 @@ import globalReducer from '../containers/App/reducer';
 import adminReducer from '../containers/Admin/reducer';
 import languageProviderReducer from '../components/LanguageProvider/reducer';
 import permissionsManagerReducer from '../components/PermissionsManager/reducer';
-import menuReducer from '../containers/LeftMenu/reducer';
+import menuReducer from '../components/LeftMenu/reducer';
 
 // TODO move containers reducers into this folder
